Move `Stroke` to its own file
