+++ wheatley/row_generation/plain_hunt_generator.py
@@ -1,6 +1,7 @@
 """ A module to hold a row generator that produces plain hunt on any (even) stage. """
 
-from wheatley.aliases import Places, Row, Stroke
+from wheatley.aliases import Places, Row
+from wheatley.stroke import Stroke
 
 from .row_generator import RowGenerator
